+++ tests/test_edgecases.py
@@ -1,7 +1,7 @@
-"""This module contains tests for edgecases.
+"""This module contains tests for edgecases."""
 
-"""
 import copy
+import sys
 
 import pytest
 import sphinx.errors
@@ -9,7 +9,8 @@ from sphinx.transforms.post_transforms import ReferencesResolver
 
 from sphinxcontrib.autodoc_pydantic import PydanticModelDocumenter
 from tests.compatibility import rst_alias_class_directive, \
-    TYPEHINTS_PREFIX, TYPING_MODULE_PREFIX_V2, module_doc_string_tab
+    TYPEHINTS_PREFIX, TYPING_MODULE_PREFIX_V2, module_doc_string_tab, \
+    PYTHON_LT_310
 
 
 def test_not_json_compliant(autodocument):
@@ -367,16 +368,30 @@ def test_autodoc_pydantic_model_show_validator_summary_inherited_with_inherited(
     assert result == actual
 
 
-def test_autodoc_pydantic_model_show_validator_summary_inherited_without_inherited(
-        autodocument):
-    """Ensure that references to inherited validators point to parent class
-    when `inherited-members` is not given.
+def test_autodoc_pydantic_model_show_validator_summary_inherited_without_inherited_no_field(
+        autodocument
+        ):
+    """Special edge case where inherited class without fields shows parent 
+    fields/validator even though `inherited-members` is not activated. 
+    This only occurs for python < 3.10.
 
     Relates to #122.
 
     """
 
     result = [
+        '',
+        '.. py:pydantic_model:: ModelShowValidatorsSummaryInherited',
+        '   :module: target.configuration',
+        '',
+        '   ModelShowValidatorsSummaryInherited.',
+        '',
+        '   :Validators:',
+        '      - :py:obj:`check_inherited <target.configuration.ModelShowValidatorsSummaryInherited.check_inherited>` » :py:obj:`field <target.configuration.ModelShowValidatorsSummaryInherited.field>`',
+        ''
+    ]
+
+    result_python_lt_310 = [
         '',
         '.. py:pydantic_model:: ModelShowValidatorsSummaryInherited',
         '   :module: target.configuration',
@@ -395,6 +410,42 @@ def test_autodoc_pydantic_model_show_validator_summary_inherited_without_inherit
         options_app={"autodoc_pydantic_model_show_validator_summary": True,
                      "autodoc_pydantic_model_members": True},
         deactivate_all=True)
+    
+    if PYTHON_LT_310:
+        assert result_python_lt_310 == actual
+    else:
+        assert result == actual
+
+
+def test_autodoc_pydantic_model_show_validator_summary_inherited_without_inherited_with_field(
+        autodocument
+        ):
+    """Ensure that references to inherited validators point to parent class
+    when `inherited-members` is not given.
+
+    Relates to #122.
+
+    """
+
+    result = [
+        '',
+        '.. py:pydantic_model:: ModelShowValidatorsSummaryInheritedWithField',
+        '   :module: target.configuration',
+        '',
+        '   ModelShowValidatorsSummaryInheritedWithField.',
+        '',
+        '   :Validators:',
+        '      - :py:obj:`check_inherited <target.configuration.ModelShowValidatorsSummaryInheritedWithField.check_inherited>` » :py:obj:`field <target.configuration.ModelShowValidatorsSummaryInheritedWithField.field>`',
+        ''
+    ]
+
+    actual = autodocument(
+        documenter='pydantic_model',
+        object_path='target.configuration.ModelShowValidatorsSummaryInheritedWithField',
+        options_app={"autodoc_pydantic_model_show_validator_summary": True,
+                     "autodoc_pydantic_model_members": True},
+        deactivate_all=True)
+    
     assert result == actual
 
 
@@ -711,4 +762,311 @@ def test_autodoc_pydantic_model_hide_reused_validator_true_identical_names(
                      "members": None,
                      "undoc-members": None},
         **kwargs)
+    assert result == actual
+
+
+def test_autodoc_pydantic_model_inherited_members_enabled_without_base_model(
+    autodocument,
+):
+    """Ensure that inheritance from parent class is correct for fields and
+    validators considering both members and summary sections.
+
+    """
+
+    kwargs = dict(
+        object_path="target.edgecase_inherited_members.enabled_without_base_model.Child",
+        documenter=PydanticModelDocumenter.objtype,
+        deactivate_all=True,
+    )
+
+    result = [
+        "",
+        ".. py:pydantic_model:: Child",
+        "   :module: target.edgecase_inherited_members.enabled_without_base_model",
+        "",
+        "   Child",
+        "",
+        "   :Fields:",
+        "      - :py:obj:`field_on_child (str) <target.edgecase_inherited_members.enabled_without_base_model.Child.field_on_child>`",
+        "      - :py:obj:`field_on_parent (str) <target.edgecase_inherited_members.enabled_without_base_model.Child.field_on_parent>`",
+        "",
+        "   :Validators:",
+        "      - :py:obj:`validate_field_on_child <target.edgecase_inherited_members.enabled_without_base_model.Child.validate_field_on_child>` » :py:obj:`field_on_child <target.edgecase_inherited_members.enabled_without_base_model.Child.field_on_child>`",
+        "      - :py:obj:`validate_field_on_parent <target.edgecase_inherited_members.enabled_without_base_model.Child.validate_field_on_parent>` » :py:obj:`field_on_parent <target.edgecase_inherited_members.enabled_without_base_model.Child.field_on_parent>`",
+        "",
+        "",
+        "   .. py:pydantic_field:: Child.field_on_child",
+        "      :module: target.edgecase_inherited_members.enabled_without_base_model",
+        "      :type: str",
+        "",
+        "      field_on_child",
+        "",
+        "",
+        "   .. py:pydantic_field:: Child.field_on_parent",
+        "      :module: target.edgecase_inherited_members.enabled_without_base_model",
+        "      :type: str",
+        "",
+        "      field_on_parent",
+        "",
+        "",
+        "   .. py:pydantic_validator:: Child.validate_field_on_child",
+        "      :module: target.edgecase_inherited_members.enabled_without_base_model",
+        "      :classmethod:",
+        "",
+        "      Validate field_on_child",
+        "",
+        "",
+        "   .. py:pydantic_validator:: Child.validate_field_on_parent",
+        "      :module: target.edgecase_inherited_members.enabled_without_base_model",
+        "      :classmethod:",
+        "",
+        "      Validate field_on_parent",
+        "",
+    ]
+
+    actual = autodocument(
+        options_app={
+            "autodoc_pydantic_model_show_validator_members": True,
+            "autodoc_pydantic_model_show_validator_summary": True,
+            "autodoc_pydantic_model_show_field_summary": True,
+        },
+        options_doc={
+            "members": None,
+            "inherited-members": "BaseModel",
+        },
+        **kwargs,
+    )
+    assert result == actual
+
+
+def test_autodoc_pydantic_model_inherited_members_disabled(autodocument):
+    """Ensure that inheritance from parent class is correct for fields and
+    validators considering both members and summary sections.
+
+    """
+
+    result = [
+        "",
+        ".. py:pydantic_model:: Child",
+        "   :module: target.edgecase_inherited_members.disabled",
+        "",
+        "   Child",
+        "",
+        "   :Fields:",
+        "      - :py:obj:`field_on_child (str) <target.edgecase_inherited_members.disabled.Child.field_on_child>`",
+        "",
+        "   :Validators:",
+        "      - :py:obj:`validate_field_on_child <target.edgecase_inherited_members.disabled.Child.validate_field_on_child>` » :py:obj:`field_on_child <target.edgecase_inherited_members.disabled.Child.field_on_child>`",
+        "",
+        "",
+        "   .. py:pydantic_field:: Child.field_on_child",
+        "      :module: target.edgecase_inherited_members.disabled",
+        "      :type: str",
+        "",
+        "      field_on_child",
+        "",
+        "",
+        "   .. py:pydantic_validator:: Child.validate_field_on_child",
+        "      :module: target.edgecase_inherited_members.disabled",
+        "      :classmethod:",
+        "",
+        "      Validate field_on_child",
+        "",
+    ]
+
+    actual = autodocument(
+        options_app={
+            "autodoc_pydantic_model_show_validator_members": True,
+            "autodoc_pydantic_model_show_validator_summary": True,
+            "autodoc_pydantic_model_show_field_summary": True,
+            "autodoc_pydantic_model_members": True,
+        },
+        deactivate_all=True,
+        object_path="target.edgecase_inherited_members.disabled.Child",
+        documenter=PydanticModelDocumenter.objtype,
+    )
+
+    assert result == actual
+
+
+def test_autodoc_pydantic_model_inherited_members_disabled_with_overwrite(autodocument):
+    """Ensure reference to parent validator is correctly set in
+    child field. Moreover, ensure that overwritten field is displayed and
+    not excluded.
+
+    """
+
+    result = [
+        "",
+        ".. py:pydantic_model:: ChildWithOverwrite",
+        "   :module: target.edgecase_inherited_members.disabled_with_overwrite",
+        "",
+        "   ChildWithOverwrite",
+        "",
+        "   :Fields:",
+        "      - :py:obj:`field_on_child (str) <target.edgecase_inherited_members.disabled_with_overwrite.ChildWithOverwrite.field_on_child>`",
+        "      - :py:obj:`field_on_parent (str) <target.edgecase_inherited_members.disabled_with_overwrite.ChildWithOverwrite.field_on_parent>`",
+        "",
+        "   :Validators:",
+        "      - :py:obj:`validate_field_on_child <target.edgecase_inherited_members.disabled_with_overwrite.ChildWithOverwrite.validate_field_on_child>` » :py:obj:`field_on_child <target.edgecase_inherited_members.disabled_with_overwrite.ChildWithOverwrite.field_on_child>`",
+        "      - :py:obj:`validate_field_on_parent <target.edgecase_inherited_members.disabled_with_overwrite.Parent.validate_field_on_parent>` » :py:obj:`field_on_parent <target.edgecase_inherited_members.disabled_with_overwrite.ChildWithOverwrite.field_on_parent>`",
+        "",
+        "",
+        "   .. py:pydantic_field:: ChildWithOverwrite.field_on_child",
+        "      :module: target.edgecase_inherited_members.disabled_with_overwrite",
+        "      :type: str",
+        "",
+        "      field_on_child",
+        "",
+        "",
+        "   .. py:pydantic_field:: ChildWithOverwrite.field_on_parent",
+        "      :module: target.edgecase_inherited_members.disabled_with_overwrite",
+        "      :type: str",
+        "",
+        "      overwritten field_on_parent",
+        "",
+        "",
+        "   .. py:pydantic_validator:: ChildWithOverwrite.validate_field_on_child",
+        "      :module: target.edgecase_inherited_members.disabled_with_overwrite",
+        "      :classmethod:",
+        "",
+        "      Validate field_on_child",
+        "",
+    ]
+
+    actual = autodocument(
+        options_app={
+            "autodoc_pydantic_model_show_validator_members": True,
+            "autodoc_pydantic_model_show_validator_summary": True,
+            "autodoc_pydantic_model_show_field_summary": True,
+            "autodoc_pydantic_model_members": True,
+        },
+        deactivate_all=True,
+        object_path="target.edgecase_inherited_members.disabled_with_overwrite.ChildWithOverwrite",
+        documenter=PydanticModelDocumenter.objtype,
+    )
+
+    assert result == actual
+
+
+def test_autodoc_module_inherited_members_disabled_with_overwrite(autodocument):
+    """Ensure that inheritance from parent class is correct for fields and
+    validators considering both members and summary sections, given that
+    a child class overwrites parent field.
+
+    Additionally, ensure reference to parent validator is correctly set in
+    childs fields. Moreover, ensure that overwritten field is displayed.
+
+    """
+
+    kwargs = dict(
+        object_path="target.edgecase_inherited_members.disabled_with_overwrite_module",
+        documenter="module",
+        deactivate_all=True,
+    )
+
+    result = [
+        "",
+        ".. py:module:: target.edgecase_inherited_members.disabled_with_overwrite_module",
+        "",
+        "",
+        ".. py:pydantic_model:: Child",
+        "   :module: target.edgecase_inherited_members.disabled_with_overwrite_module",
+        "",
+        "   Child",
+        "",
+        "   :Fields:",
+        "      - :py:obj:`field_on_child (str) <target.edgecase_inherited_members.disabled_with_overwrite_module.Child.field_on_child>`",
+        "",
+        "   :Validators:",
+        "      - :py:obj:`validate_field_on_child <target.edgecase_inherited_members.disabled_with_overwrite_module.Child.validate_field_on_child>` » :py:obj:`field_on_child <target.edgecase_inherited_members.disabled_with_overwrite_module.Child.field_on_child>`",
+        "",
+        "",
+        "   .. py:pydantic_field:: Child.field_on_child",
+        "      :module: target.edgecase_inherited_members.disabled_with_overwrite_module",
+        "      :type: str",
+        "",
+        "      field_on_child",
+        "",
+        "",
+        "   .. py:pydantic_validator:: Child.validate_field_on_child",
+        "      :module: target.edgecase_inherited_members.disabled_with_overwrite_module",
+        "      :classmethod:",
+        "",
+        "      Validate field_on_child",
+        "",
+        "",
+        ".. py:pydantic_model:: ChildWithOverwrite",
+        "   :module: target.edgecase_inherited_members.disabled_with_overwrite_module",
+        "",
+        "   ChildWithOverwrite",
+        "",
+        "   :Fields:",
+        "      - :py:obj:`field_on_child (str) <target.edgecase_inherited_members.disabled_with_overwrite_module.ChildWithOverwrite.field_on_child>`",
+        "      - :py:obj:`field_on_parent (str) <target.edgecase_inherited_members.disabled_with_overwrite_module.ChildWithOverwrite.field_on_parent>`",
+        "",
+        "   :Validators:",
+        "      - :py:obj:`validate_field_on_child <target.edgecase_inherited_members.disabled_with_overwrite_module.ChildWithOverwrite.validate_field_on_child>` » :py:obj:`field_on_child <target.edgecase_inherited_members.disabled_with_overwrite_module.ChildWithOverwrite.field_on_child>`",
+        "      - :py:obj:`validate_field_on_parent <target.edgecase_inherited_members.disabled_with_overwrite_module.Parent.validate_field_on_parent>` » :py:obj:`field_on_parent <target.edgecase_inherited_members.disabled_with_overwrite_module.ChildWithOverwrite.field_on_parent>`",
+        "",
+        "",
+        "   .. py:pydantic_field:: ChildWithOverwrite.field_on_child",
+        "      :module: target.edgecase_inherited_members.disabled_with_overwrite_module",
+        "      :type: str",
+        "",
+        "      field_on_child",
+        "",
+        "",
+        "   .. py:pydantic_field:: ChildWithOverwrite.field_on_parent",
+        "      :module: target.edgecase_inherited_members.disabled_with_overwrite_module",
+        "      :type: str",
+        "",
+        "      overwritten field_on_parent",
+        "",
+        "",
+        "   .. py:pydantic_validator:: ChildWithOverwrite.validate_field_on_child",
+        "      :module: target.edgecase_inherited_members.disabled_with_overwrite_module",
+        "      :classmethod:",
+        "",
+        "      Validate field_on_child",
+        "",
+        "",
+        ".. py:pydantic_model:: Parent",
+        "   :module: target.edgecase_inherited_members.disabled_with_overwrite_module",
+        "",
+        "   Base",
+        "",
+        "   :Fields:",
+        "      - :py:obj:`field_on_parent (str) <target.edgecase_inherited_members.disabled_with_overwrite_module.Parent.field_on_parent>`",
+        "",
+        "   :Validators:",
+        "      - :py:obj:`validate_field_on_parent <target.edgecase_inherited_members.disabled_with_overwrite_module.Parent.validate_field_on_parent>` » :py:obj:`field_on_parent <target.edgecase_inherited_members.disabled_with_overwrite_module.Parent.field_on_parent>`",
+        "",
+        "",
+        "   .. py:pydantic_field:: Parent.field_on_parent",
+        "      :module: target.edgecase_inherited_members.disabled_with_overwrite_module",
+        "      :type: str",
+        "",
+        "      field_on_parent",
+        "",
+        "",
+        "   .. py:pydantic_validator:: Parent.validate_field_on_parent",
+        "      :module: target.edgecase_inherited_members.disabled_with_overwrite_module",
+        "      :classmethod:",
+        "",
+        "      Validate field_on_parent",
+        "",
+    ]
+
+    actual = autodocument(
+        options_app={
+            "autodoc_pydantic_model_show_validator_members": True,
+            "autodoc_pydantic_model_show_validator_summary": True,
+            "autodoc_pydantic_model_show_field_summary": True,
+        },
+        options_doc={
+            "members": None,
+        },
+        **kwargs,
+    )
     assert result == actual